(fix) Fixed objection js returning string instead of int

+++ src/database/migrations/20200411005018_create_agent_table.ts
@@ -3,7 +3,7 @@ import * as Knex from "knex";
 export async function up(knex: Knex): Promise<any> {
   return knex.schema.createTable("agent", (table) => {
     table
-      .bigInteger("userId")
+      .integer("userId")
       .unsigned()
       .index()
       .references("id")
